pyproject.toml
[project]
name = "localai-backend"
version = "0.1.0"
description = "LocalAI backend server"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate==1.9.0",
    "aiohappyeyeballs==2.6.1",
    "aiohttp==3.12.14",
    "aiosignal==1.4.0",
    "alabaster==1.0.0",
    "alembic==1.16.4",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "asttokens==3.0.0",
    "attr==0.3.2",
    "attrs==25.3.0",
    "babel==2.17.0",
    "basicsr-fixed>=1.4.2",
    "bidict==0.23.1",
    "brotli==1.2.0",
    "brotlicffi==1.1.0.0",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "chardet==5.2.0",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorlog==6.9.0",
    "comm==0.2.3",
    "configparser==7.2.0",
    "coverage==7.10.4",
    "cryptography==45.0.5",
    "decorator==5.2.1",
    "diffusers==0.35.2",
    "docopt==0.6.2",
    "docutils==0.21.2",
    "executing==2.2.0",
    "fanstatic==1.5",
    "fastapi==0.116.1",
    "filelock==3.18.0",
    "frozenlist==1.7.0",
    "fsspec==2025.7.0",
    "greenlet==3.2.4",
    "h11==0.16.0",
    "h2==4.2.0",
    "hf-xet==1.2.0",
    "hpack==4.1.0",
    "htmlparser==0.0.2",
    "httpcore==1.0.9",
    "httpx==0.27.2",
    "huggingface-hub==0.34.0",
    "hyperframe==6.1.0",
    "idna==3.10",
    "imagesize==1.4.1",
    "importlib-metadata==8.7.0",
    "iniconfig==2.1.0",
    "ipython==9.4.0",
    "ipython-pygments-lexers==1.1.1",
    "ipywidgets==8.1.7",
    "jaraco-classes==3.4.0",
    "jaraco-context==6.0.1",
    "jaraco-functools==4.3.0",
    "jedi==0.19.2",
    "jeepney==0.9.0",
    "jinja2==3.1.6",
    "js==1.0",
    "jupyterlab-widgets==3.0.15",
    "keyring==25.6.0",
    "legacy-cgi==2.6.3",
    "mako==1.3.10",
    "markupsafe==3.0.2",
    "matplotlib-inline==0.1.7",
    "more-itertools==10.7.0",
    "mpmath==1.3.0",
    "multidict==6.6.4",
    "networkx==3.5",
    "numpy==2.3.2",
    "packaging==25.0",
    "parso==0.8.4",
    "peft>=0.18.0",
    "pexpect==4.9.0",
    "pillow==11.3.0",
    "pip==25.3",
    "pipreqs==0.4.13",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.51",
    "propcache==0.3.2",
    "protobuf>=6.33.1",
    "psutil==7.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "pydash>=8.0.5",
    "pygments==2.19.2",
    "pyopenssl==25.1.0",
    "pypdl>=1.5.7",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "python-engineio==4.12.2",
    "python-socketio==5.14.0",
    "pyyaml==6.0.2",
    "realesrgan>=0.3.0",
    "redis==6.2.0",
    "regex==2025.7.34",
    "requests==2.32.5",
    "roman-numerals-py==3.1.0",
    "safetensors==0.6.2",
    "scipy>=1.16.3",
    "secretstorage==3.3.3",
    "setuptools==80.9.0",
    "simple-websocket==1.1.0",
    "simplejson==3.20.1",
    "sniffio==1.3.1",
    "snowballstemmer==3.0.1",
    "sphinx==8.2.3",
    "sphinxcontrib-applehelp==2.0.0",
    "sphinxcontrib-devhelp==2.0.0",
    "sphinxcontrib-htmlhelp==2.1.0",
    "sphinxcontrib-jsmath==1.0.1",
    "sphinxcontrib-qthelp==2.0.0",
    "sphinxcontrib-serializinghtml==2.0.0",
    "sqlalchemy==2.0.41",
    "stack-data==0.6.3",
    "starlette==0.47.2",
    "sympy==1.13.1",
    "tenacity==9.1.2",
    "thread==2.0.5",
    "tokenizers==0.22.1",
    "torch==2.6.0+cu124",
    "torchsde>=0.2.6",
    "torchvision==0.21.0",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "trampoline>=0.1.2",
    "transformers==4.57.1",
    "triton==3.2.0; sys_platform == 'linux'",
    "ty>=0.0.1a28",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "urllib3-secure-extra==0.1.0",
    "uvicorn==0.35.0",
    "wcwidth==0.2.13",
    "webob==1.8.9",
    "widgetsnbextension==4.0.14",
    "wsproto==1.2.0",
    "yarg==0.1.10",
    "yarl==1.20.1",
    "zipp==3.23.0",
    "zstandard==0.23.0",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app"]

[tool.setuptools.package-data]
app = ["py.typed"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
addopts = "-ra --strict-markers"
markers = ["slow: slow tests", "integration: integration tests"]

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "alembic/*", "static/*", "*/__pycache__/*"]
relative_files = true
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
    "raise NotImplementedError",
]

[tool.coverage.xml]
output = "coverage.xml"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[dependency-groups]
dev = [
    "httpx==0.27.2",
    "pytest==8.3.3",
    "pytest-asyncio==0.24.0",
    "pytest-cov==5.0.0",
    "ty>=0.0.1a28",
    "types-pillow>=10.2.0.20240822",
    "types-psutil>=7.0.0.20251116",
    "types-pyyaml>=6.0.12.20250915",
    "types-redis>=4.6.0.20241004",
    "types-requests>=2.32.4.20250913",
    "types-setuptools>=80.9.0.20250822",
    "types-tqdm>=4.67.0.20250809",
    "types-ujson>=5.10.0.20250822",
]
